Extract reusable UI components to reduce duplication
Extract four shared components from duplicated code across 17 page files:
- EnvironmentTab: key-value editor for environment variables
- ToolListEditor: tool pattern list with add/remove and @-prefix styling
- ValidationProblemsPanel: severity-based validation message display
- ConfirmDeleteModal: standardized destructive action confirmation dialog

Net reduction of ~550 lines across the UI module.
