+++ ui/src/pages/SecretsPage.tsx
@@ -28,6 +28,7 @@ import {
     updateSecret,
     deleteSecret,
 } from "../config/api";
+import { ConfirmDeleteModal } from "../components/ConfirmDeleteModal";
 
 export function SecretsPage() {
     const [secrets, setSecrets] = useState<Secret[]>([]);
@@ -144,20 +145,10 @@ export function SecretsPage() {
                 )}
             </div>
 
-            <Modal isOpen={deleteTarget !== null} onClose={() => setDeleteTarget(null)} variant="small">
-                <ModalHeader title="Delete Secret" />
-                <ModalBody>
-                    Delete this secret? This cannot be undone.
-                </ModalBody>
-                <ModalFooter>
-                    <Button variant="danger" onClick={confirmDelete}>
-                        Delete
-                    </Button>
-                    <Button variant="link" onClick={() => setDeleteTarget(null)}>
-                        Cancel
-                    </Button>
-                </ModalFooter>
-            </Modal>
+            <ConfirmDeleteModal isOpen={deleteTarget !== null} title="Delete Secret"
+                onConfirm={confirmDelete} onCancel={() => setDeleteTarget(null)}>
+                Delete this secret? This cannot be undone.
+            </ConfirmDeleteModal>
 
             <Modal isOpen={isModalOpen} onClose={() => setIsModalOpen(false)} variant="medium">
                 <ModalHeader title={editing ? "Update Secret" : "Add Secret"} />
