+++ ui/src/pages/ToolDetailPage.tsx
@@ -51,6 +51,7 @@ import ExclamationCircleIcon from "@patternfly/react-icons/dist/esm/icons/exclam
 import { EditLabelsModal } from "../components/EditLabelsModal";
 import { LabelDisplay } from "../components/LabelDisplay";
 import { ToolAiModal } from "../components/ToolAiModal";
+import { ValidationProblemsPanel } from "../components/ValidationProblemsPanel";
 
 export function ToolDetailPage() {
     const { toolId } = useParams<{ toolId: string }>();
@@ -241,33 +242,7 @@ export function ToolDetailPage() {
                         </TabTitleText>
                     }>
                         <TabContent id="problems-tab" eventKey={4} activeKey={activeTab} style={{ marginTop: "24px" }}>
-                            <div style={{ maxWidth: "700px" }}>
-                                {validationMessages.map((msg, i) => (
-                                    <div key={i} style={{
-                                        display: "flex",
-                                        alignItems: "flex-start",
-                                        gap: 8,
-                                        padding: "10px 12px",
-                                        marginBottom: 4,
-                                        borderRadius: 4,
-                                        backgroundColor: msg.severity === "error"
-                                            ? "#fef3f2" : "#fdf7e7",
-                                        border: `1px solid ${msg.severity === "error"
-                                            ? "#c9190b" : "#f0ab00"}`,
-                                    }}>
-                                        {msg.severity === "error"
-                                            ? <ExclamationCircleIcon style={{ color: "#c9190b", marginTop: 2 }} />
-                                            : <ExclamationTriangleIcon style={{ color: "#f0ab00", marginTop: 2 }} />
-                                        }
-                                        <div>
-                                            <div style={{ fontSize: "13px" }}>{msg.message}</div>
-                                            <div style={{ fontSize: "12px", color: "#6a6e73", marginTop: 2 }}>
-                                                Field: <code>{msg.field}</code>
-                                            </div>
-                                        </div>
-                                    </div>
-                                ))}
-                            </div>
+                            <ValidationProblemsPanel messages={validationMessages} />
                         </TabContent>
                     </Tab>
                 )}
