+++ ui/src/pages/ToolsetDetailPage.tsx
@@ -15,9 +15,8 @@ import {
     TextInput,
     Title,
 } from "@patternfly/react-core";
-import { AddToolInput } from "../components/AddToolInput";
+import { ToolListEditor } from "../components/ToolListEditor";
 import SaveIcon from "@patternfly/react-icons/dist/esm/icons/save-icon";
-import TimesIcon from "@patternfly/react-icons/dist/esm/icons/times-icon";
 import {
     type Toolset,
     type NewToolset,
@@ -135,63 +134,27 @@ export function ToolsetDetailPage() {
                 </FormGroup>
             </Form>
 
-            <div style={{ maxWidth: "700px" }}>
-                <Title headingLevel="h2" size="md" style={{ marginBottom: "8px" }}>
-                    Tools ({tools.length})
-                </Title>
-                <p style={{ color: "#6a6e73", marginBottom: "16px" }}>
+            <Title headingLevel="h2" size="md" style={{ marginBottom: "8px" }}>
+                Tools ({tools.length})
+            </Title>
+            <ToolListEditor
+                tools={tools}
+                onAdd={addTool}
+                onRemove={removeTool}
+                onReplace={replaceTools}
+                helpText={<>
                     Define the tools in this toolset. You can add individual tool
                     patterns like <code>Bash(git log *)</code> or reference other
                     toolsets using the <code>@ToolsetName</code> syntax.
-                </p>
-
-                <div style={{ marginBottom: "16px" }}>
-                    <AddToolInput onAdd={addTool} onReplace={replaceTools} existingTools={tools} />
-                </div>
-
-                {tools.length === 0 ? (
+                </>}
+                emptyContent={
                     <EmptyState variant="xs">
                         <EmptyStateBody>
                             No tools in this toolset yet.
                         </EmptyStateBody>
                     </EmptyState>
-                ) : (
-                    <div style={{ display: "flex", flexDirection: "column", gap: "4px" }}>
-                        {tools.map((tool) => (
-                            <Flex key={tool}
-                                alignItems={{ default: "alignItemsCenter" }}
-                                style={{
-                                    padding: "8px 12px",
-                                    backgroundColor: tool.startsWith("@")
-                                        ? "var(--pf-t--global--background--color--primary--default)"
-                                        : "var(--pf-t--global--background--color--secondary--default)",
-                                    borderRadius: "4px",
-                                    border: tool.startsWith("@")
-                                        ? "1px solid var(--pf-t--global--border--color--default)"
-                                        : "none",
-                                }}>
-                                <FlexItem grow={{ default: "grow" }}>
-                                    <code style={{
-                                        fontSize: "13px",
-                                        color: tool.startsWith("@")
-                                            ? "var(--pf-t--global--color--brand--default)"
-                                            : "inherit",
-                                    }}>
-                                        {tool}
-                                    </code>
-                                </FlexItem>
-                                <FlexItem>
-                                    <Button variant="plain" size="sm"
-                                        onClick={() => removeTool(tool)}
-                                        aria-label={`Remove ${tool}`}>
-                                        <TimesIcon />
-                                    </Button>
-                                </FlexItem>
-                            </Flex>
-                        ))}
-                    </div>
-                )}
-            </div>
+                }
+            />
         </PageSection>
     );
 }
